pyproject.toml
[tool.poetry]
name = "recontee"
version = "0.1.0"
description = "All-in-one web recon pipeline"
authors = ["juniotee"]
packages = [{ include = "recontee", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.12.3"
rich = "^13.7.1"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
recontee = "recontee.cli:app"
